+++ src/main/java/com/example/pro/repository/TimeSlotRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TimeSlotRepository extends CrudRepository<TimeSlot, Integer> {
+public interface TimeSlotRepository extends CrudRepository<TimeSlot, Long> {
     List<TimeSlot> findByPractitionerId(int practitionerId);
     List<TimeSlot> findByPractitionerIdAndStatusIn(int practitionerId, List<TimeSlotStatus> statuses);
 }
